server/requirements.txt
kafka-python~=2.0.2
rasa~=3.0.6
spacy~=3.2.1
nltk~=3.6.7
fastapi~=0.73.0
uvicorn~=0.17.1